+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 #### 27.11.2018
 
+* [VideoSearch] Better jumping to position after loading more videos
 * [VideoSearch] Further scrolling with a bottom progress bar
 
 #### 26.11.2018